+++ frontend-material/src/assets/scss/base/_colors.scss
@@ -1,4 +1,4 @@
-$primary-green: #077915;
+$primary-green: #9c27b0;
 $primary-orange: #fc7710;
 $white: #fff;
 $gray-eighty: #ccc;
@@ -10,4 +10,4 @@ $red: #e23d3d;
 $gray-light-bg: #f5f5f5;
 $gray-medium-bg: #f1f1f1;
 $gray-dark-bg: #dddddd;
-$green-light-bg: rgba(147, 220, 156, 0.15);
+$green-light-bg: #9c27b0;
