+++ frontend-material/src/assets/scss/components/_index.scss
@@ -2,4 +2,3 @@
 @import "products";
 @import "product";
 @import "footer";
-@import "modal";
