+++ frontend-material/src/assets/scss/components/_modal.scss
