+++ frontend-material/src/empty-states/EmptyCart.js
@@ -1,13 +1,14 @@
 import React, { Component } from "react";
+import empty_cart from "assets/img/empty-cart.png"
 
 const EmptyCart = props => {
   return (
     <div className="empty-cart">
       <img
-        src="https://res.cloudinary.com/sivadass/image/upload/v1495427934/icons/empty-cart.png"
+        src={empty_cart}
         alt="empty-cart"
       />
-      <h2>You cart is empty!</h2>
+      <h2>Your cart is empty!</h2>
     </div>
   );
 };
