+++ frontend-material/src/empty-states/NoResults.js
@@ -1,15 +1,17 @@
 import React, { Component } from "react";
+import tree from "assets/img/bare-tree.png"
+
 
 const NoResults = () => {
   return (
     <div className="products">
       <div className="no-results">
         <img
-          src="https://res.cloudinary.com/sivadass/image/upload/v1494699523/icons/bare-tree.png"
+          src={tree}
           alt="Empty Tree"
         />
         <h2>Sorry, no products matched your search!</h2>
-        <p>Enter a different keyword and try.</p>
+        <p>Enter a different keyword and try again.</p>
       </div>
     </div>
   );
